requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
django-cors-headers==4.6.0
fastapi==0.115.8
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==13.1
whitenoise==6.8.2
yarl==1.18.3
